Renaming to nixops_gcp and updating README
