+++ default.nix
@@ -3,7 +3,6 @@ let
   overrides = import ./overrides.nix { inherit pkgs; };
 in pkgs.poetry2nix.mkPoetryApplication {
   projectDir = ./.;
-  python = pkgs.python3;
   overrides = pkgs.poetry2nix.overrides.withDefaults overrides;
   meta.description = "Nix package for ${pkgs.stdenv.system}";
 }
