+++ nixops_gcp/gcp_common.py
