+++ nixops_gcp/nix/common-gce-options.nix
