+++ nixops_gcp/nix/gce-credentials.nix
