+++ nixops_gcp/nix/gce-disk.nix
