+++ nixops_gcp/nix/gce-forwarding-rule.nix
