+++ nixops_gcp/nix/gce-http-health-check.nix
