+++ nixops_gcp/nix/gce-image.nix
