+++ nixops_gcp/nix/gce-images.nix
