+++ nixops_gcp/nix/gce-static-ip.nix
