+++ nixops_gcp/nix/gse-bucket.nix
