+++ nixops_gcp/resources/gce_network.py
@@ -8,7 +8,7 @@ from libcloud.compute.types import Provider
 from libcloud.compute.providers import get_driver
 
 from nixops.util import attr_property
-from nixops-gcp.gcp_common import ResourceDefinition, ResourceState
+from nixops_gcp.gcp_common import ResourceDefinition, ResourceState
 
 
 def normalize_list(tags):
