+++ nixops_gcp/resources/gce_route.py
@@ -6,7 +6,7 @@ import libcloud.common.google
 
 from nixops import backends
 from nixops.util import attr_property
-from nixops-gcp.gcp_common import ResourceDefinition, ResourceState
+from nixops_gcp.gcp_common import ResourceDefinition, ResourceState
 
 
 class GCERouteDefinition(ResourceDefinition):
