+++ nixops_gcp/resources/gce_target_pool.py
@@ -5,9 +5,9 @@
 import os
 import libcloud.common.google
 
-from nixops-gcp.resources.gce_http_health_check import GCEHTTPHealthCheckState
+from nixops_gcp.resources.gce_http_health_check import GCEHTTPHealthCheckState
 from nixops.util import attr_property
-from nixops-gcp.gcp_common import ResourceDefinition, ResourceState, optional_string
+from nixops_gcp.gcp_common import ResourceDefinition, ResourceState, optional_string
 
 class GCETargetPoolDefinition(ResourceDefinition):
     """Definition of a GCE Target Pool"""
