+++ pyproject.toml
@@ -1,10 +1,10 @@
 [tool.poetry]
-name = "nixops-gcp"
+name = "nixops_gcp"
 version = "1.0"
 description = "NixOps backend for Google Cloud Platform"
 authors = ["Amine Chikhaoui <amine.chikhaoui91@gmail.com>"]
 license = "MIT"
-include = [ "nixops-gcp/nix/*.nix" ]
+include = [ "nixops_gcp/nix/*.nix" ]
 
 [tool.poetry.dependencies]
 python = "^3.7"
@@ -12,12 +12,12 @@ apache-libcloud = "^3.0.0"
 nixops = {git = "https://github.com/NixOS/nixops.git", rev = "master"}
 
 [tool.poetry.dev-dependencies]
-nose = "^1"
+nose = "^1.3.7"
 mypy = "^0.770"
 black = "^19.10b0"
 
 [tool.poetry.plugins."nixops"]
-gcp = "nixops-gcp.plugin"
+gcp = "nixops_gcp.plugin"
 
 [build-system]
 requires = ["poetry>=0.12"]
